Stop using index.jsx & rename folder
